chemistry: fix issues from cpplint, refs #439

+++ src/chemistry/reactions/aqueous_equilibrium_complex.cc
@@ -13,19 +13,20 @@ AqueousEquilibriumComplex::AqueousEquilibriumComplex()
 }  // end AqueousEquilibriumComplex() constructor
 
 
-AqueousEquilibriumComplex::AqueousEquilibriumComplex(const SpeciesName name,
-                                                     const SpeciesId id,
-                                                     const std::vector<SpeciesName>& species,
-                                                     const std::vector<double>& stoichiometry,
-                                                     const std::vector<SpeciesId>& species_ids,
-                                                     const double h2o_stoich,
-                                                     const double charge,
-                                                     const double mol_wt,
-                                                     const double size,
-                                                     const double logK)
-    : SecondarySpecies(name, id, species, stoichiometry, species_ids, h2o_stoich, charge, mol_wt, size, logK)
-
-{
+AqueousEquilibriumComplex::AqueousEquilibriumComplex(
+    const SpeciesName name,
+    const SpeciesId id,
+    const std::vector<SpeciesName>& species,
+    const std::vector<double>& stoichiometry,
+    const std::vector<SpeciesId>& species_ids,
+    const double h2o_stoich,
+    const double charge,
+    const double mol_wt,
+    const double size,
+    const double logK)
+    : SecondarySpecies(name, id, species, stoichiometry,
+                       species_ids, h2o_stoich, charge,
+                       mol_wt, size, logK) {
 }  // end AqueousEquilibriumComplex() constructor
 
 AqueousEquilibriumComplex::~AqueousEquilibriumComplex() {
@@ -40,7 +41,8 @@ void AqueousEquilibriumComplex::Update(const std::vector<Species>& primarySpecie
   */
   double lnQK = -lnK();
   for (int i = 0; i < ncomp(); i++) {
-    lnQK += stoichiometry_.at(i) * primarySpecies.at(species_ids_.at(i)).ln_activity();
+    lnQK += stoichiometry_.at(i) *
+        primarySpecies.at(species_ids_.at(i)).ln_activity();
   }
   lnQK_ = lnQK;
   //  molality_ = std::exp(lnQK) / act_coef_;
@@ -60,7 +62,7 @@ void AqueousEquilibriumComplex::AddContributionToDTotal(
   // taking derivative of contribution to residual in row i with respect
   // to species in column j
 
-  // TODO: because of memory layout, c loops should be for(i){for(j)}...?
+  // TODO(bandre): because of memory layout, c loops should be for(i){for(j)}...?
 
   // column loop
   std::cout.precision(15);
@@ -78,13 +80,13 @@ void AqueousEquilibriumComplex::AddContributionToDTotal(
 
 void AqueousEquilibriumComplex::display(void) const {
   std::cout << "    " << name() << " = ";
-  // TODO: uncomment and update test output
+  // TODO(bandre): uncomment and update test output
   //   if (h2o_stoichiometry_ > 0) {
   //     std::cout << h2o_stoichiometry_ << " " << "H2O" << " + ";
   //   }
-  for (int i = 0; i < (int)species_names_.size(); i++) {
+  for (unsigned int i = 0; i < species_names_.size(); i++) {
     std::cout << stoichiometry_.at(i) << " " << species_names_.at(i);
-    if (i < (int)species_names_.size() - 1) {
+    if (i < species_names_.size() - 1) {
       std::cout << " + ";
     }
   }
@@ -97,13 +99,13 @@ void AqueousEquilibriumComplex::display(void) const {
 void AqueousEquilibriumComplex::Display(void) const {
   std::cout << "    " << name() << " = "
             << std::fixed << std::setprecision(3);
-  // TODO: uncomment and update test output
+  // TODO(bandre): uncomment and update test output
   //   if (h2o_stoichiometry_ > 0) {
   //     std::cout << h2o_stoichiometry_ << " " << "H2O" << " + ";
   //   }
-  for (int i = 0; i < (int)species_names_.size(); i++) {
+  for (unsigned int i = 0; i < species_names_.size(); i++) {
     std::cout << stoichiometry_.at(i) << " " << species_names_.at(i);
-    if (i < (int)species_names_.size() - 1) {
+    if (i < species_names_.size() - 1) {
       std::cout << " + ";
     }
   }
